Added a check to ask user to delete docker images

+++ tests/e2e/local/minikube/setup_test.sh
@@ -4,8 +4,11 @@
 . ./setup_minikube_env.sh
 
 # Remove old imges.
-echo "Removing Old Images tagged localhost:5000/*:latest"
-docker images localhost:5000/*:latest -q | xargs docker rmi -f
+read -p "Do you want to delete old docker images tagged localhost:5000/*:latest[default: no]: " update
+delete_images=${update:-"no"}
+if [[ $delete_images = *"y"* ]] || [[ $delete_images = *"Y"* ]]; then
+  docker images localhost:5000/*:latest -q | xargs docker rmi -f
+fi
 
 # Make and Push images to insecure local registry on VM.
 # Set GOOS=linux to make sure linux binaries are built on macOS
